+++ packages/manager/src/features/VPC/VPCLanding.tsx
@@ -1,12 +1,9 @@
 import { VPC } from '@linode/api-v4/lib/vpcs/types';
 import * as React from 'react';
-import { Link } from 'react-router-dom';
 
-import { Action } from 'src/components/ActionMenu';
 import { CircleProgress } from 'src/components/CircleProgress';
 import { ErrorState } from 'src/components/ErrorState/ErrorState';
 import { Hidden } from 'src/components/Hidden';
-import { InlineMenuAction } from 'src/components/InlineMenuAction/InlineMenuAction';
 import LandingHeader from 'src/components/LandingHeader';
 import { PaginationFooter } from 'src/components/PaginationFooter/PaginationFooter';
 import { ProductInformationBanner } from 'src/components/ProductInformationBanner/ProductInformationBanner';
@@ -18,24 +15,13 @@ import { TableRow } from 'src/components/TableRow';
 import { TableSortCell } from 'src/components/TableSortCell';
 import { useOrder } from 'src/hooks/useOrder';
 import { usePagination } from 'src/hooks/usePagination';
-import { useRegionsQuery } from 'src/queries/regions';
 import { useVPCsQuery } from 'src/queries/vpcs';
 import { getAPIErrorOrDefault } from 'src/utilities/errorUtils';
 
 import { VPCEmptyState } from './VPCEmptyState';
+import { VPCRow } from './VPCRow';
 
 const preferenceKey = 'vpcs';
-
-const actions: Action[] = [
-  {
-    onClick: () => null,
-    title: 'Edit',
-  },
-  {
-    onClick: () => null,
-    title: 'Delete',
-  },
-];
 
 const VPCLanding = () => {
   const pagination = usePagination(1, preferenceKey);
@@ -59,7 +45,6 @@ const VPCLanding = () => {
     },
     filter
   );
-  const { data: regions } = useRegionsQuery();
 
   if (error) {
     return (
@@ -142,40 +127,9 @@ const VPCLanding = () => {
         </TableHead>
         <TableBody>
           {vpcs?.data.map((vpc: VPC) => {
-            const regionLabel = regions?.find((r) => r.id === vpc.region)
-              ?.label;
-            return (
-              <TableRow
-                ariaLabel={`VPC ${vpc.label}`}
-                data-qa-vpc-id={vpc.id}
-                key={`vpc-row-${vpc.id}`}
-              >
-                <TableCell>
-                  <Link to={`/vpc/${vpc.id}`}>{vpc.label}</Link>
-                </TableCell>
-                <Hidden smDown>
-                  <TableCell>{regionLabel}</TableCell>
-                </Hidden>
-                <Hidden mdDown>
-                  <TableCell>{vpc.id}</TableCell>
-                </Hidden>
-                <TableCell>{vpc.subnets.length}</TableCell>
-                <Hidden mdDown>
-                  <TableCell>{vpc.subnets.length}</TableCell>
-                </Hidden>
-                <TableCell actionCell>
-                  {actions.map((action) => {
-                    return (
-                      <InlineMenuAction
-                        actionText={action.title}
-                        key={action.title}
-                        onClick={action.onClick}
-                      />
-                    );
-                  })}
-                </TableCell>
-              </TableRow>
-            );
+              return (
+                <VPCRow vpc={vpc} key={vpc.id} />
+              )
           })}
         </TableBody>
       </Table>
