+++ packages/manager/src/features/VPC/VPCRow.tsx
@@ -1,0 +1,71 @@
+import { VPC } from '@linode/api-v4/lib/vpcs/types';
+import * as React from 'react';
+import { Link } from 'react-router-dom';
+
+import { Action } from 'src/components/ActionMenu';
+import { Hidden } from 'src/components/Hidden';
+import { InlineMenuAction } from 'src/components/InlineMenuAction/InlineMenuAction';
+import { TableCell } from 'src/components/TableCell';
+import { TableRow } from 'src/components/TableRow';
+import { useRegionsQuery } from 'src/queries/regions';
+
+interface Props {
+  vpc: VPC;
+}
+
+const actions: Action[] = [
+  {
+    onClick: () => null,
+    title: 'Edit',
+  },
+  {
+    onClick: () => null,
+    title: 'Delete',
+  },
+];
+
+export const VPCRow = ({ vpc }: Props) => {
+  const { id, label, subnets } = vpc;
+  const { data: regions } = useRegionsQuery();
+
+  const regionLabel = regions?.find((r) => r.id === vpc.region)?.label ?? '';
+  const numLinodes = subnets.reduce(
+    (acc, subnet) => acc + subnet.linodes.length,
+    0
+  );
+
+  return (
+    <TableRow
+      ariaLabel={`VPC ${label}`}
+      data-qa-vpc-id={id}
+      key={`vpc-row-${id}`}
+    >
+      <TableCell>
+        <Link to={`/vpc/${id}`}>{label}</Link>
+      </TableCell>
+      <Hidden smDown>
+        <TableCell>{regionLabel}</TableCell>
+      </Hidden>
+      <Hidden mdDown>
+        <TableCell>{id}</TableCell>
+      </Hidden>
+      <TableCell>{subnets.length}</TableCell>
+      <Hidden mdDown>
+        <TableCell>{numLinodes}</TableCell>
+      </Hidden>
+      <TableCell actionCell>
+        {actions.map((action) => {
+          return (
+            <InlineMenuAction
+              actionText={action.title}
+              key={action.title}
+              onClick={action.onClick}
+            />
+          );
+        })}
+      </TableCell>
+    </TableRow>
+  );
+};
+
+export default VPCRow;
